% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinic_Animal.R
\docType{class}
\name{Animal-class}
\alias{Animal-class}
\title{Animal-class}
\description{
Animal class
}
\section{Slots}{

\describe{
\item{\code{Time}}{Time}

\item{\code{Group}}{Group}

\item{\code{ID}}{ID}

\item{\code{weight}}{weight}

\item{\code{tumorVolume}}{tumorVolume}
}}

\seealso{
\code{\link{animal}}
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
\keyword{classes}
