% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinic_Animal.R
\name{Animal-method}
\alias{Animal-method}
\alias{as.data.frame.Animal}
\alias{plot.Animal}
\title{Animal-method}
\usage{
\method{as.data.frame}{Animal}(object)

\method{plot}{Animal}(
  object,
  type = c("tumorVolume", "weight")[1],
  list.color = {     list.color <- mycolor[1:7]     names(list.color) <- paste("G",
    1:7, sep = "")     list.color },
  legend.position = "right",
  label.size = 0.15,
  plot.size = 25,
  verbose = T
)

\S4method{statistics}{Animal}(
  object,
  type = c("tumorVolume", "weight")[1],
  targetTime = NULL,
  method = "wilcox.test",
  paired = FALSE,
  p.adjust.method = "BH",
  verbose = T
)
}
\arguments{
\item{object}{\code{\link{Animal-class}}}

\item{type}{One of \code{tumorVolume} and \code{weight}}

\item{list.color}{A color vector with names of \code{Group} slot.}

\item{legend.position}{the position of legends ("none", "left", "right",
"bottom", "top", or two-element numeric vector)}

\item{label.size}{Size of label border, in mm.}

\item{plot.size}{The size of the \code{ggplot} plot.}

\item{verbose}{Whether report process and results}

\item{targetTime}{Strings like '20220120'. Default is \code{NULL}, which means that every time would be considerated.}

\item{method}{the type of test. Default is \link[stats]{wilcox.test}. Allowed
values include: \itemize{ \item \code{\link[stats]{t.test}} (parametric) and
\code{\link[stats]{wilcox.test}} (non-parametric). Perform comparison
between two groups of samples. If the grouping variable contains more than
two levels, then a pairwise comparison is performed. \item
\code{\link[stats]{anova}} (parametric) and
\code{\link[stats]{kruskal.test}} (non-parametric). Perform one-way ANOVA
test comparing multiple groups. }}

\item{paired}{a logical indicating whether you want a paired test. Used only
in \code{\link[stats]{t.test}} and in \link[stats]{wilcox.test}.}

\item{p.adjust.method}{method for adjusting p values (see
 \code{\link[stats]{p.adjust}}). Has impact only in a situation, where
 multiple pairwise tests are performed; or when there are multiple grouping
 variables. Allowed values include "holm", "hochberg", "hommel",
 "bonferroni", "BH", "BY", "fdr", "none". If you don't want to adjust the p
 value (not recommended), use p.adjust.method = "none".

 Note that, when the \code{formula} contains multiple variables, the p-value
 adjustment is done independently for each variable.}
}
\value{
a data frame
}
\description{
\code{as.data.frame} method for \code{Animal} class

\code{plot} method for \code{Animal} class

\code{statistics} method for \code{Animal} class
}
\examples{
list.color <- mycolor[1:7]
names(list.color) <- paste('G',1:7,sep = '')
}
\seealso{
\code{\link{Animal}}
}
\author{
Weibin Huang<\email{654751191@qq.com}
}
