% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinic_Animal.R
\name{animal}
\alias{animal}
\title{Create an Animal object from a data frame}
\usage{
animal(
  x,
  Time = "Time",
  Group = "Group",
  ID = "ID",
  tumorSize = "tumor_size_mm",
  weight = "weigth_g"
)
}
\arguments{
\item{x}{a data frame}

\item{Time}{Time.Character.}

\item{Group}{The group of animals}

\item{ID}{The ID of animals}

\item{tumorSize}{a specified tumor size vector for animals. Always looks like: '7.15,8.56;4.85,9.65;'}

\item{weight}{The weight of animals}
}
\value{
\code{\link{Animal-class}}
}
\description{
Create an Animal object from a data frame
}
\author{
Weibin Huang<\email{654751191@qq.com}
}
