% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_CCK8.R
\name{FastCCK8}
\alias{FastCCK8}
\title{Fast methods for CCK8 data analysis}
\usage{
FastCCK8(
  data,
  cell.type = "cell",
  group = "group",
  empty.group = "0",
  stimulate.time = "stimulate.time",
  observe.time = "observe.time",
  value = "od",
  x.lab = "time",
  y.lab = "OD value(450nm)",
  title = "",
  size = 20,
  method.two = "t.test",
  method.all = "anova",
  verbose = T
)
}
\arguments{
\item{data}{a data frame}

\item{cell.type}{the colname of cell type}

\item{group}{the colname of group}

\item{empty.group}{the colname of the reference group without cells in the 96-well board.}

\item{stimulate.time}{the colname of a stimulating factor.For example,different dose of drugs;different treatments.}

\item{observe.time}{colname.observation time after CCK8 treatment.Time nodes like 1h,2h,4h are recommanded.}

\item{value}{colname of OD value}

\item{x.lab}{the title of x axis}

\item{y.lab}{the title of y axis}

\item{title}{the title of plot}

\item{size}{the size of plot}

\item{verbose}{whether do report}
}
\value{
a \code{LuckyCCK8} object
}
\description{
Fast methods for CCK8 data analysis
}
\examples{
data <- readxl::read_xlsx("E:/iProjects/exosome/experiment/exosome_1A/CCK8/CCK8_HGC27的合适种植数量及观察时间_20190424.xlsx", na = "NA")

object <- lucky::FastCCK8(data)
View(object$Data$statistics$two)
View(object$Data$statistics$all)

select <- summary(object,observe_time = "2h")
}
\seealso{
\code{\link{plot}}.
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
