% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinic_getCoxPie.R
\name{getCoxPie}
\alias{getCoxPie}
\title{Get pie plot of Cox chisq proportion for variants in Cox regression model}
\usage{
getCoxPie(
  data,
  time.col = "time",
  status.col = "event",
  clusters,
  time.raw.type = c("Day", "Month", "Year")[1],
  time.target.type = c("Day", "Month", "Year")[3],
  global = T,
  pie.color,
  Others.color = mycolor[27],
  pie.label,
  dc = 1,
  legend.name = "Variable",
  legend.position = "right",
  is.title = T,
  plot.size = 7,
  names = "project"
)
}
\arguments{
\item{data}{a data frame with survival data(time & event)}

\item{time.col}{colname of time value}

\item{status.col}{Numeric.\code{status.col} must be a binary parameter with 0(event not happened yet) or 1(event happened) value.}

\item{clusters}{list of characters.Some of the colnames of the data representing the value involving the glm model.}

\item{time.raw.type}{one of "Day","Month","Year"}

\item{time.target.type}{one of "Day","Month","Year"}

\item{global}{
    should a global chi-square test be done, in addition to the 
    per-variable or per-term tests tests. 
  }

\item{pie.color}{a character vector with color value and variable names. Default is missing.}

\item{Others.color}{the color of others. It works when \code{global=T}}

\item{dc}{the digit counts of proportion}

\item{legend.name}{the name of legened}

\item{legend.position}{the position of legend}

\item{is.title}{whether plot title}

\item{plot.size}{the size of saved plots}

\item{names}{part of save files or plots.}
}
\value{
pie data and plots
}
\description{
Get pie plot of Cox chisq proportion for variants in Cox regression model
}
\examples{
data("myeloma",package = "survminer");
data <- myeloma;rm(myeloma)
show.color(mycolor)
pie.color <- c(TP53="#FB8072", WHSC1="#BEBADA", DEPDC1="#80B1D3", Others="#A65628")
pie.label = c("TP53" = "tp53","WHSC1" = "whsc1","DEPDC1"="depdc1")
a <- getCoxPie(data,
               time.col = "time",
               status.col = "event",
               clusters = list(c("TP53","WHSC1"),
                               c("TP53","WHSC1","DEPDC1")),
               time.raw.type = c("Day", "Month","Year")[2],
               time.target.type = c("Day", "Month", "Year")[3],
               global=T,
               pie.color = pie.color,
               pie.label = pie.label,
               plot.size = 7,
               names = "project")
}
\seealso{
\code{\link[survival]{cox.zph}}; \code{\link[survival]{cox.ph}}
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
