% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinic_getCoxTable.R
\name{getCoxTable}
\alias{getCoxTable}
\title{Get a model result of univariate and multivariate Cox regression Models}
\usage{
getCoxTable(
  data,
  time.col,
  status.col,
  is.multiple = T,
  stepwise = F,
  direction = "both",
  cluster,
  control,
  dig = 2,
  names = "test1"
)
}
\arguments{
\item{data}{a data frame containing cluster,time and status value}

\item{time.col}{colnames of time value}

\item{status.col}{colnames of status value}

\item{is.multiple}{whether to do multiple cox regression}

\item{stepwise}{Whether use stepwise strategy to simple multiple cox model}

\item{direction}{method of stepwise.See \code{\link[MASS]{stepAIC}}.}

\item{cluster}{the cluster values like age and sex}

\item{control}{set the control value in a factor vector.For example, in T status,T1 should be a control in the contrast of T3 vs T1.}

\item{dig}{the decimal place of output numeric value.}

\item{names}{part of file name.}
}
\description{
Get a model result of univariate and multivariate Cox regression Models
}
\seealso{
\code{\link[MASS]{stepAIC}}
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
