% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinic_getDCA2.R
\name{getDCA2}
\alias{getDCA2}
\title{ggplot style for decision estimate analysis}
\usage{
getDCA2(
  data,
  group.by = NULL,
  time.col = "time",
  status.col = "event",
  clusters,
  time.raw.type = c("Day", "Month", "Year")[1],
  time.target.type = c("Day", "Month", "Year")[3],
  time.knot = c(3, 5),
  study.design = "cohort",
  population.prevalence = c(NA, NA),
  load.existed.res = F,
  plot.ntbft.by = c("group", "model")[1],
  plot.ntbft.ymin = NULL,
  plot.ntbft.ymax = NULL,
  width = 9,
  height = 7,
  add.bootc = T,
  add.cv = F,
  add.ci = T,
  plot.legend.position = "right",
  plot.type = c("time", "cluster", "disperse")[1],
  plot.label,
  parameter.label = c(all = "All positive", none = "All negative"),
  plot.family = "Arial",
  names = "project"
)
}
\arguments{
\item{data}{a data frame with survival data(time & event)}

\item{group.by}{group colname}

\item{time.col}{colname of time value}

\item{status.col}{Numeric.\code{status.col} must be a binary parameter with 0(event not happened yet) or 1(event happened) value.}

\item{clusters}{list of characters.Some of the colnames of the data representing the value involving the glm model.}

\item{time.raw.type}{one of "Day","Month","Year"}

\item{time.target.type}{one of "Day","Month","Year"}

\item{time.knot}{time knot to create survival probability}

\item{study.design}{Either 'cohort' (default) or 'case-control' describing the study design used to obtain data. See details for more information.}

\item{population.prevalence}{Outcome prevalence rate in the population used to calculate decision curves when study.design = 'case-control'.}

\item{load.existed.res}{whether to use existed rds result in the work space}

\item{plot.ntbft.ymin}{the ymin of plot.ntbft. Default is \code{ymin=-0.05}}

\item{plot.ntbft.ymax}{the ymax of plot.ntbft. Default is \code{ymax=max(nb[,c(nolines,nobands)],na.rm=T)}}

\item{plot.legend.position}{the position of plot legend}

\item{plot.type}{Character. One/some of c("time","cluster","disperse"). if \code{plot.type="time"},then multiple clusters would be ploted for every time; if \code{plot.type="cluster"}, then multiple times would be ploted for every cluster; if \code{plot.type="disperse"}, then every single plot would be ploted indivitually.}

\item{plot.label}{List. The name of list are the paste0 of elements in one of \code{clusters} with "-".The content of the list are the real names that would be printed in the plot legend/title.}

\item{parameter.label}{the label of parameter label.}

\item{names}{part of save files or plots.}
}
\value{
ggplot style plot and result
}
\description{
ggplot style for decision estimate analysis
}
\details{
beta version
}
\examples{
## This is a simulative process and available only with CORRECT VARIABLES
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
