% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinic_getDCA2Compare.R
\name{getDCA2Compare}
\alias{getDCA2Compare}
\title{Comparision of getDCA2 result}
\usage{
getDCA2Compare(
  object,
  glop_start = 0.01,
  glop_end = 0.1,
  glop_step = 0.01,
  seed = c(127, 128),
  names = "project"
)
}
\arguments{
\item{object}{the result of \code{\link{getDCA2}}}

\item{glop_start}{the start of threshold probability}

\item{glop_end}{the end of threshold probability}

\item{glop_step}{the step of threshold probability}

\item{seed}{Numeric with 2 values. The first seed is for the boot of \code{nbdiff}, and the second one is for the boot of \code{areadiff}.}

\item{names}{part of save files or plots.}
}
\description{
Comparision of getDCA2 result
}
\seealso{
\code{\link{getDCA2}}
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
