% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{LuckyCCK8-class}
\alias{LuckyCCK8-class}
\title{Class "LuckyCCK8"}
\description{
Class "LuckyCCK8"
}
\section{Slots}{

\describe{
\item{\code{Repeat}}{repeat parameters}

\item{\code{Data}}{datasets}

\item{\code{Plot}}{ggplot}
}}

\author{
Weibin Huang<\email{654751191@qq.com}>
}
\keyword{classes}
