% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_CCK8.R
\docType{class}
\name{luckyExperiment-classes}
\alias{luckyExperiment-classes}
\title{LuckyCCK8-class}
\description{
LuckyCCK8-class
}
\section{Slots}{

\describe{
\item{\code{Repeat}}{repeat parameters}

\item{\code{Data}}{datasets}

\item{\code{Plot}}{ggplot}
}}

\author{
Weibin Huang<\email{654751191@qq.com}>
}
\keyword{classes}
