% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_CCK8.R
\docType{methods}
\name{summary}
\alias{summary}
\title{summary method}
\usage{
\S4method{summary}{LuckyCCK8}(
  object,
  observe_time = "2h",
  x.lab = "time",
  y.lab = "OD value(450nm)",
  title = "",
  size = 20,
  legend.position = "bottom",
  verbose = T
)
}
\arguments{
\item{x.lab}{the title of x axis}

\item{y.lab}{the title of y axis}

\item{title}{the title of plot}

\item{size}{the size of plot}

\item{verbose}{whether do report}
}
\value{
New plot of CCK8 result
}
\description{
plot method for LuckyCCK8 Object
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
