% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinic_Animal.R
\name{tumor_volume}
\alias{tumor_volume}
\title{Create an Animal object from a data frame}
\usage{
tumor_volume(x, tumorSize = "tumor_size_mm")
}
\arguments{
\item{x}{a data frame}

\item{tumorSize}{a specified tumor size vector for animals. Always looks like: '7.15,8.56;4.85,9.65;'}
}
\value{
Tumor volume for \code{\link{Animal-class}}
}
\description{
Create an Animal object from a data frame
}
\author{
Weibin Huang<\email{654751191@qq.com}
}
